added stack monitoring and appropriate test

+++ src/Auth/Authenticator.c
@@ -197,11 +197,11 @@ int8_t AuthCopyAndDecryptVerify(uint8_t key[], uint8_t key_len)
 		return AUTH_ERR_INVALID_PTR;
 	}
 
-	uint8_t  *dst = &_sauth;
+	uint32_t  *dst = &_sauth;
 
     size_t section_len = (size_t)(&_eauth - &_sauth);
 
-	uint8_t * src = &_sloadauth;
+	uint32_t * src = &_sloadauth;
 
     memcpy(dst, src, section_len);
